requirements.txt
adjustText
alembic
altair
anthropic
cattrs
datasets
fastapi==0.100.0
fire
hydra-core
matplotlib
mistralai
openai==0.28.0
pandas
pebble
polars
pre-commit
pydantic
replicate
scikit-learn
scipy
seaborn
sqlalchemy==2.0.18
tabulate
tenacity
termcolor
tiktoken
tqdm
trueskill
typer[all]
uvicorn[standard]==0.22.0
wandb
